% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_datasets.R
\name{analyze_dataset}
\alias{analyze_dataset}
\title{analyze_dataset: Basically performs preprocessing and then returns analyzed RNASeq dataset (diff. expression)
          , i.e. the DESeq2 result whose p-values and baseMean statistics can then be used with IHW}
\usage{
analyze_dataset(dataset = c("airway", "bottomly", "hammer"), res = TRUE)
}
\arguments{
\item{dataset}{Character, name of dataset to be preprocessed, only 3 choices currently available ("airway","bottomly","hammer")}

\item{res}{(default TRUE): return result table, rather than DESeq2 object}
}
\value{
Preprocessed dataset
}
\description{
analyze_dataset: Basically performs preprocessing and then returns analyzed RNASeq dataset (diff. expression)
          , i.e. the DESeq2 result whose p-values and baseMean statistics can then be used with IHW
}
\examples{
bottomly <- analyze_dataset("bottomly")

}
