% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_methods.R
\name{bh}
\alias{bh}
\title{bh: Wrapper for Benjamini Hochberg}
\usage{
bh(unadj_p, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}
}
\value{
BH multiple testing object
}
\description{
bh: Wrapper for Benjamini Hochberg
}
\examples{
     sim_df <- du_ttest_sim(20000,0.95, 1.5)
     obj <- bh(sim_df$pvalue, .1)
     sum(rejected_hypotheses(obj))

}
