% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_methods.R
\name{bonf}
\alias{bonf}
\title{bonf: Wrapper for Bonferroni}
\usage{
bonf(unadj_p, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}
}
\value{
Bonferroni multiple testing object
}
\description{
bonf: Wrapper for Bonferroni
}
\examples{
     sim_df <- du_ttest_sim(20000,0.95, 1.5)
     obj <- bonf(sim_df$pvalue, .1)
     sum(rejected_hypotheses(obj))

}
