% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{cai}
\alias{cai}
\title{cai: Cai's local fdr based method}
\usage{
cai(unadj_p, groups, alpha, lfdr_estimation = "fdrtool")
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}

\item{lfdr_estimation}{Method used to estimate the loca fdr, defaults to fdrtool}
}
\description{
cai: Cai's local fdr based method
}

