% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulations.R
\name{null_sim}
\alias{null_sim}
\alias{null_sim_fun}
\title{Null simulation: Generate uniformly distributed p-values and covariates}
\usage{
null_sim(m, seed = NULL)

null_sim_fun(m)
}
\arguments{
\item{m}{Integer, total number of hypotheses}

\item{seed}{Integer, Random seed to be used for simulation 
(default: NULL, i.e. RNG state will be used as is)}
}
\value{
A data frame containing all information about the simulation experiment
}
\description{
Null simulation: Generate uniformly distributed p-values and covariates
}
\section{Functions}{
\itemize{
\item \code{null_sim_fun}: Creates a closure function for a given seed
}}

