% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pretty_legend}
\alias{pretty_legend}
\title{helper function to create nice legends}
\usage{
pretty_legend(gg, last_vals, xmin, fontsize = 13)
}
\arguments{
\item{gg}{ggplot2 object}

\item{last_vals, }{data frame with columns 
label, colour, last_vals (i.e. place label with colour at y-coordinate last_vals)}

\item{xmin}{Numeric, offset}

\item{fontsize}{Integer, fontsize}
}
\value{
Another ggplot2 object

   http://www.r-bloggers.com/coloring-and-drawing-outside-the-lines-in-ggplot/
}
\description{
helper function to create nice legends
}

