% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_methods.R
\name{storey_qvalue}
\alias{storey_qvalue}
\title{storey_qvalue: Wrapper for Storey's qvalue package}
\usage{
storey_qvalue(unadj_p, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}
}
\value{
StoreyQValue multiple testing object
}
\description{
storey_qvalue: Wrapper for Storey's qvalue package
}
\examples{
     sim_df <- du_ttest_sim(20000,0.95, 1.5)
     obj <- storey_qvalue(sim_df$pvalue, .1)
     sum(rejected_hypotheses(obj))

}
