% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{tst_gbh}
\alias{tst_gbh}
\title{tst_gbh: wrapper for gbh with method="TST"
lsl_gbh: wrapper for gbh with method="LSL"}
\usage{
tst_gbh(unadj_p, groups, alpha, ...)
}
\description{
tst_gbh: wrapper for gbh with method="TST"
lsl_gbh: wrapper for gbh with method="LSL"
}

