% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/analyze_datasets.R
\name{analyze_dataset}
\alias{analyze_dataset}
\title{analyze_dataset: Basically performs preprocessing and then returns analyzed RNASeq dataset (diff. expression)
          , i.e. the DESeq2 result whose p-values and baseMean statistics can then be used with DDHW}
\usage{
analyze_dataset(dataset = c("pasilla", "airway", "bottomly", "pasilla"),
  res = T)
}
\arguments{
\item{dataset}{Character, name of dataset to be preprocessed, only 4 choices currently available}

\item{res}{(default TRUE): return result table, rather than DESeq2 object}
}
\description{
analyze_dataset: Basically performs preprocessing and then returns analyzed RNASeq dataset (diff. expression)
          , i.e. the DESeq2 result whose p-values and baseMean statistics can then be used with DDHW
}

