% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stratified_methods.R
\name{gbh}
\alias{gbh}
\title{gbh: Grouped Benjamini Hochberg}
\usage{
gbh(unadj_p, groups, alpha, method = "TST", pi0_global = "weighted_average")
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{groups}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}

\item{method}{What pi0 estimator should be used (available "TST","LSL")}

\item{pi0_global}{Method requires also a pi0 estimate for the marginal p-value distribution. Can either apply pi0 estimation method to
all p-values (pi0_global="global") or use a weigted averarage (pi0_global="weighted_average") of the pi0 estimates within each stratum.
This is not explicitly stated in the paper, but based on a reproduction of their paper figures it seems to be the weighted_average.}
}
\description{
gbh: Grouped Benjamini Hochberg
}

