% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi0_estimators.R
\name{lsl_pi0_est}
\alias{lsl_pi0_est}
\title{LSL (Least-Slope) pi0 estimator}
\usage{
lsl_pi0_est(pvalue)
}
\arguments{
\item{pvalue}{Numeric vector of unadjusted p-values.}
}
\value{
estimated proportion of null hypotheses (pi0)
}
\description{
LSL (Least-Slope) pi0 estimator
}
\examples{
     sim_df <- du_ttest_sim(20000,0.95, 1.5)
     lsl_pi0_est(sim_df$pvalue)

}
