% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/covariate_methods.R
\name{scott_fdrreg}
\alias{scott_fdrreg}
\title{scott_fdrreg: Wrapper for FDR regression}
\usage{
scott_fdrreg(unadj_p, filterstat, alpha, df = 3)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{filterstat}{Factor to which different hypotheses belong}

\item{alpha}{Significance level at which to apply method}
}
\description{
scott_fdrreg: Wrapper for FDR regression
}

