% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_methods.R
\name{storey_qvalue}
\alias{storey_qvalue}
\title{storey_qvalue: Wrapper for Storey's qvalue package}
\usage{
storey_qvalue(unadj_p, alpha)
}
\arguments{
\item{unadj_p}{Numeric vector of unadjusted p-values.}

\item{alpha}{Significance level at which to apply method}
}
\description{
storey_qvalue: Wrapper for Storey's qvalue package
}

